compact microsoft_craziness.h into find_msvc.c
